requirements.txt
sentence-transformers
faiss-cpu
numpy
tqdm
openai
google-generativeai
flask
